Updating Workers REST API
