+++ crowdsourcing/serializers/requester.py
@@ -1,9 +1,7 @@
-__author__ = 'elsabakiu'
-from crowdsourcing import models
-from datetime import datetime
-from rest_framework import serializers
-import json
+__author__ = ['dmorina', 'asmita', 'megha']
 
+from crowdsourcing import models
+from rest_framework import serializers
 
 class RequesterSerializer(serializers.ModelSerializer):
     class Meta:
@@ -13,3 +11,10 @@ class RequesterSerializer(serializers.ModelSerializer):
 class RequesterRankingSerializer(serializers.ModelSerializer):
     class Meta:
         model = models.RequesterRanking
+
+
+class QualificationSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = models.Qualification
+        fields = ('module', 'types', 'type', 'created_timestamp', 'last_updated')
+        read_only_fields = ('module', 'types', 'created_timestamp', 'last_updated')
