+++ crowdsourcing/serializers/task.py
@@ -1,27 +1,17 @@
-__author__ = 'elsabakiu'
+__author__ = ['elsabakiu', 'asmita', 'megha']
 
 from crowdsourcing import models
-from datetime import datetime
 from rest_framework import serializers
-import json
-
-
-class TaskPropertiesSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = models.Task
-
-
-class TaskWorkerSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = models.TaskWorker
-        
-
-class QualificationSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = models.Qualification
-        
 
 class TaskSerializer(serializers.ModelSerializer):
     class Meta:
         model = models.Task        
-        fields = ( 'module', 'statuses','status', 'deleted', 'created_timestamp')
+        fields = ('module', 'statuses', 'status', 'deleted', 'created_timestamp', 'last_updated')
+        read_only_fields = ('module', 'statuses', 'created_timestamp', 'last_updated')
+
+
+class CurrencySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = models.Currency
+        fields = ('name', 'iso_code', 'last_updated')
+        read_only_fields = ('last_updated')
