+++ crowdsourcing/serializers/worker.py
@@ -1,14 +1,11 @@
-__author__ = ['elsabakiu', 'dmorina', 'neilthemathguy', 'megha']
+__author__ = ['elsabakiu', 'dmorina', 'neilthemathguy', 'megha', 'asmita']
 
 from crowdsourcing import models
-from datetime import datetime
 from rest_framework import serializers
-import json
 
 class WorkerSerializer(serializers.ModelSerializer):
     class Meta:
         model = models.Worker
-        
 
 
 class SkillSerializer(serializers.ModelSerializer):
@@ -44,14 +41,3 @@ class WorkerModuleApplicationSerializer (serializers.ModelSerializer):
         model = models.WorkerModuleApplication
         fields = ('worker', 'module', 'status', 'created_timestamp', 'last_updated')
         read_only_fields = ('worker', 'module', 'created_timestamp', 'last_updated')
-        
-        
-class CurrencySerializer(serializers.ModelSerializer):
-    class Meta:
-        model = models.Currency
-        
-
-        
-
-
-
