+++ crowdsourcing/viewsets/requester.py
@@ -1,4 +1,5 @@
-__author__ = 'dmorina'
+__author__ = ['dmorina', 'asmita', 'megha']
+
 from crowdsourcing.serializers.requester import *
 from rest_framework import viewsets
 
@@ -7,7 +8,14 @@ class RequesterViewSet(viewsets.ModelViewSet):
     queryset = Requester.objects.all()
     serializer_class = RequesterSerializer
 
+
 class RequesterRankingViewSet(viewsets.ModelViewSet):
     from crowdsourcing.models import RequesterRanking
     queryset = RequesterRanking.objects.all()
     serializer_class = RequesterRankingSerializer
+
+
+class QualificationViewSet(viewsets.ModelViewSet):
+    from crowdsourcing.models import Qualification
+    queryset = Qualification.objects.all()
+    serializer_class = QualificationSerializer
