+++ crowdsourcing/viewsets/task.py
@@ -1,16 +1,15 @@
+__author__ = ['elsabakiu', 'asmita', 'megha']
 
 from crowdsourcing.serializers.task import *
-from rest_framework import generics
-from rest_framework import status, views as rest_framework_views, viewsets
-
+from rest_framework import viewsets
 
 class TaskViewSet(viewsets.ModelViewSet):
     from crowdsourcing.models import Task
     queryset = Task.objects.all()
     serializer_class = TaskSerializer
-    
 
-class QualificationViewSet(viewsets.ModelViewSet):
-    from crowdsourcing.models import Qualification
-    queryset = Qualification.objects.all()
-    serializer_class = QualificationSerializer
+
+class CurrencyViewSet(viewsets.ModelViewSet):
+    from crowdsourcing.models import Currency
+    queryset = Currency.objects.all()
+    serializer_class = CurrencySerializer
