+++ fixtures/meghaWorkerData.csv
@@ -220,5 +220,5 @@ worker,module,status,created_timestamp,last_updated
 0,0,1,2015-05-01T00:17:40.085Z,2015-05-01T00:17:40.085Z
 1,1,2,2015-05-01T00:17:40.085Z,2015-05-01T00:17:40.085Z
 2,2,3,2015-05-01T00:17:40.085Z,2015-05-01T00:17:40.085Z
-3,3,4,2015-05-01T00:17:40.085Z,2015-05-01T00:17:40.085Z
+3,3,2,2015-05-01T00:17:40.085Z,2015-05-01T00:17:40.085Z
 4,4,1,2015-05-01T00:17:40.085Z,2015-05-01T00:17:40.085Z
